requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.3
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector==2.2.9
Werkzeug==3.1.8
zipp==3.23.1
Flask-Login
mysql-connector-python
